+++ imars_etl/extract.py
@@ -1,14 +1,9 @@
 import os
 import logging
 
-from imars_etl.object_storage.ObjectStorageHandler import ObjectStorageHandler
-from imars_etl.object_storage.ObjectStorageHandler \
-    import DEFAULT_OBJ_STORE_CONN_ID
-from imars_etl.metadata_db.MetadataDBHandler import MetadataDBHandler
-from imars_etl.metadata_db.MetadataDBHandler import DEFAULT_METADATA_DB_CONN_ID
-from imars_etl.config_logger import config_logger
-from imars_etl.exceptions.TooManyMetadataMatchesException \
-    import TooManyMetadataMatchesException
+from imars_etl.object_storage.imars_objects import imars_objects
+from imars_etl.util.config_logger import config_logger
+from imars_etl.metadata_db.mysql import meta_db_select
 
 EXTRACT_DEFAULTS = {
 }
@@ -25,8 +20,6 @@ def extract(
     sql,
     output_path=None,
     first=False,
-    metadata_conn_id=DEFAULT_METADATA_DB_CONN_ID,
-    object_store=DEFAULT_OBJ_STORE_CONN_ID,
     verbose=0,
     method=EXTRACT_METHOD.COPY[0],
     **kwargs
@@ -42,36 +35,7 @@ def extract(
 
     full_sql_str = "SELECT filepath FROM file WHERE {}".format(sql)
 
-    metadata_db = MetadataDBHandler(
-        metadata_db=metadata_conn_id,
-    )
-    try:
-        result = metadata_db.get_records(
-            full_sql_str,
-            first=first,
-        )
-    except TooManyMetadataMatchesException:
-        # === handle acceptable duplicate multihash entries (issue #41)
-        if first:  # shouldn't get here if first == True anyway
-            raise
-        multihash_sql = full_sql_str.replace(
-            "SELECT filepath FROM file ",
-            "SELECT multihash FROM file "
-        )
-        multihashes = metadata_db.get_records(
-            multihash_sql,
-            check_result=False
-        )
-        # if all list elements are equal
-        if multihashes.count(multihashes[0]) == len(multihashes):
-            # we must re-query here bc first result was stoppered by exception
-            result = metadata_db.get_records(
-                full_sql_str,
-                first=True,
-            )
-
-    src_path = result[0][0]
-
+    src_path = meta_db_select(full_sql_str)
     if output_path is None:
         output_path = "./" + os.path.basename(src_path)
 
@@ -84,14 +48,7 @@ def extract(
         os.symlink(src_path, output_path)  # ln -s src_path output_path
         fpath = output_path
     elif method.lower() in EXTRACT_METHOD.COPY:
-        object_storage = ObjectStorageHandler(
-            sql=sql,
-            output_path=output_path,
-            first=first,
-            metadata_conn_id=metadata_conn_id,
-            object_store=object_store,
-            **kwargs
-        )
+        object_storage = imars_objects()
         # use connection to download & then print a path to where the file can
         # be accessed on the local machine.
         fpath = object_storage.extract(